requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.3
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
